+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 Unreleased
 ----------
 
+### Added
+
+* `setCaching(boolean)` method. This caches the Bridge state and stops querying the Bridge
+  unless explicitly told to do that by using the `refresh()` method, or until caching is
+  disabled again.
+
 v1.1.0 (2019-05-12)
 -------------------